requirements.txt
google-genai>=1.0.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
redis>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

